move grouping method to Assembly object, make assemblyRules Seq[_] rather than T[Seq[_]]
