+++ main/src/mill/modules/Assembly.scala
@@ -1,9 +1,14 @@
 package mill.modules
 
+import java.io.InputStream
 import java.util.jar.JarFile
 import java.util.regex.Pattern
 
-import upickle.default.{ReadWriter => RW}
+import ammonite.ops._
+import geny.Generator
+import mill.Agg
+
+import scala.collection.JavaConverters._
 
 object Assembly {
 
@@ -30,16 +35,93 @@ object Assembly {
       def apply(pattern: String): ExcludePattern = ExcludePattern(Pattern.compile(pattern))
     }
     case class ExcludePattern(pattern: Pattern) extends Rule
+  }
 
-    implicit lazy val appendRW: RW[Append] = upickle.default.macroRW
-    implicit lazy val appendPatternRW: RW[AppendPattern] = upickle.default.macroRW
-    implicit lazy val excludeRW: RW[Exclude] = upickle.default.macroRW
-    implicit lazy val excludePatternRW: RW[ExcludePattern] = upickle.default.macroRW
-    implicit lazy val regexRW: RW[Pattern] = upickle.default.readwriter[String].bimap[Pattern](
-      _.pattern(), Pattern.compile
-    )
-    implicit lazy val rulesFormat: RW[Rule] = {
-      RW.merge(appendRW, appendPatternRW, excludeRW, excludePatternRW)
+  def groupAssemblyEntries(inputPaths: Agg[Path], assemblyRules: Seq[Assembly.Rule]): Map[String, GroupedEntry] = {
+    val rulesMap = assemblyRules.collect {
+      case r@Rule.Append(path) => path -> r
+      case r@Rule.Exclude(path) => path -> r
+    }.toMap
+
+    val appendPatterns = assemblyRules.collect {
+      case Rule.AppendPattern(pattern) => pattern.asPredicate().test(_)
+    }
+
+    val excludePatterns = assemblyRules.collect {
+      case Rule.ExcludePattern(pattern) => pattern.asPredicate().test(_)
+    }
+
+    classpathIterator(inputPaths).foldLeft(Map.empty[String, GroupedEntry]) {
+      case (entries, entry) =>
+        val mapping = entry.mapping
+
+        rulesMap.get(mapping) match {
+          case Some(_: Assembly.Rule.Exclude) =>
+            entries
+          case Some(_: Assembly.Rule.Append) =>
+            val newEntry = entries.getOrElse(mapping, AppendEntry.empty).append(entry)
+            entries + (mapping -> newEntry)
+
+          case _ if excludePatterns.exists(_(mapping)) =>
+            entries
+          case _ if appendPatterns.exists(_(mapping)) =>
+            val newEntry = entries.getOrElse(mapping, AppendEntry.empty).append(entry)
+            entries + (mapping -> newEntry)
+
+          case _ if !entries.contains(mapping) =>
+            entries + (mapping -> WriteOnceEntry(entry))
+          case _ =>
+            entries
+        }
     }
   }
+
+  private def classpathIterator(inputPaths: Agg[Path]): Generator[AssemblyEntry] = {
+    Generator.from(inputPaths)
+      .filter(exists)
+      .flatMap {
+        p =>
+          if (p.isFile) {
+            val jf = new JarFile(p.toIO)
+            Generator.from(
+              for(entry <- jf.entries().asScala if !entry.isDirectory)
+                yield JarFileEntry(entry.getName, () => jf.getInputStream(entry))
+            )
+          }
+          else {
+            ls.rec.iter(p)
+              .filter(_.isFile)
+              .map(sub => PathEntry(sub.relativeTo(p).toString, sub))
+          }
+      }
+  }
+}
+
+private[modules] sealed trait GroupedEntry {
+  def append(entry: AssemblyEntry): GroupedEntry
+}
+
+private[modules] object AppendEntry {
+  val empty: AppendEntry = AppendEntry(Nil)
+}
+
+private[modules] case class AppendEntry(entries: List[AssemblyEntry]) extends GroupedEntry {
+  def append(entry: AssemblyEntry): GroupedEntry = copy(entries = entry :: this.entries)
+}
+
+private[modules] case class WriteOnceEntry(entry: AssemblyEntry) extends GroupedEntry {
+  def append(entry: AssemblyEntry): GroupedEntry = this
+}
+
+private[this] sealed trait AssemblyEntry {
+  def mapping: String
+  def inputStream: InputStream
+}
+
+private[this] case class PathEntry(mapping: String, path: Path) extends AssemblyEntry {
+  def inputStream: InputStream = read.getInputStream(path)
+}
+
+private[this] case class JarFileEntry(mapping: String, getIs: () => InputStream) extends AssemblyEntry {
+  def inputStream: InputStream = getIs()
 }
